fix: address maintainer review — remove callback indirection, relocate tests
- Remove showErrorOverlay callback from surfaceMissingNodes; return
  boolean so callers invoke overlay themselves (simpler, no indirection)
- Move missingNodesErrorStore tests to co-located test file
- Fix false-positive E2E assertion: use page-scope for not.toBeVisible
  check when overlay may be removed from DOM entirely
