+++ src/stores/executionErrorStore.test.ts
@@ -36,204 +36,6 @@ vi.mock(
 import { useExecutionErrorStore } from './executionErrorStore'
 import { useMissingNodesErrorStore } from '@/platform/nodeReplacement/missingNodesErrorStore'
 
-describe('executionErrorStore — missing node operations', () => {
-  beforeEach(() => {
-    setActivePinia(createPinia())
-  })
-
-  describe('setMissingNodeTypes', () => {
-    it('sets missingNodesError with provided types', () => {
-      const store = useMissingNodesErrorStore()
-      const types: MissingNodeType[] = [
-        { type: 'NodeA', nodeId: '1', isReplaceable: false }
-      ]
-      store.setMissingNodeTypes(types)
-
-      expect(store.missingNodesError).not.toBeNull()
-      expect(store.missingNodesError?.nodeTypes).toHaveLength(1)
-      expect(store.hasMissingNodes).toBe(true)
-    })
-
-    it('clears missingNodesError when given empty array', () => {
-      const store = useMissingNodesErrorStore()
-      store.setMissingNodeTypes([
-        { type: 'NodeA', nodeId: '1', isReplaceable: false }
-      ])
-      expect(store.missingNodesError).not.toBeNull()
-
-      store.setMissingNodeTypes([])
-      expect(store.missingNodesError).toBeNull()
-      expect(store.hasMissingNodes).toBe(false)
-    })
-
-    it('deduplicates string entries by value', () => {
-      const store = useMissingNodesErrorStore()
-      store.setMissingNodeTypes([
-        'NodeA',
-        'NodeA',
-        'NodeB'
-      ] as MissingNodeType[])
-
-      expect(store.missingNodesError?.nodeTypes).toHaveLength(2)
-    })
-
-    it('deduplicates object entries by nodeId when present', () => {
-      const store = useMissingNodesErrorStore()
-      store.setMissingNodeTypes([
-        { type: 'NodeA', nodeId: '1', isReplaceable: false },
-        { type: 'NodeA', nodeId: '1', isReplaceable: false },
-        { type: 'NodeA', nodeId: '2', isReplaceable: false }
-      ])
-
-      // Same nodeId='1' deduplicated, nodeId='2' kept
-      expect(store.missingNodesError?.nodeTypes).toHaveLength(2)
-    })
-
-    it('deduplicates object entries by type when nodeId is absent', () => {
-      const store = useMissingNodesErrorStore()
-      store.setMissingNodeTypes([
-        { type: 'NodeA', isReplaceable: false },
-        { type: 'NodeA', isReplaceable: true }
-      ] as MissingNodeType[])
-
-      // Same type, no nodeId → deduplicated
-      expect(store.missingNodesError?.nodeTypes).toHaveLength(1)
-    })
-
-    it('keeps distinct nodeIds even when type is the same', () => {
-      const store = useMissingNodesErrorStore()
-      store.setMissingNodeTypes([
-        { type: 'NodeA', nodeId: '1', isReplaceable: false },
-        { type: 'NodeA', nodeId: '2', isReplaceable: false },
-        { type: 'NodeA', nodeId: '3', isReplaceable: false }
-      ])
-
-      expect(store.missingNodesError?.nodeTypes).toHaveLength(3)
-    })
-  })
-
-  describe('surfaceMissingNodes', () => {
-    beforeEach(() => {
-      mockShowErrorsTab.value = false
-    })
-
-    it('stores missing node types when called', () => {
-      const missingNodesStore = useMissingNodesErrorStore()
-      const executionErrorStore = useExecutionErrorStore()
-      const types: MissingNodeType[] = [
-        { type: 'NodeA', nodeId: '1', isReplaceable: false }
-      ]
-      missingNodesStore.surfaceMissingNodes(types, () =>
-        executionErrorStore.showErrorOverlay()
-      )
-
-      expect(missingNodesStore.missingNodesError).not.toBeNull()
-      expect(missingNodesStore.missingNodesError?.nodeTypes).toHaveLength(1)
-      expect(missingNodesStore.hasMissingNodes).toBe(true)
-    })
-
-    it('opens error overlay when ShowErrorsTab setting is true', () => {
-      mockShowErrorsTab.value = true
-      const missingNodesStore = useMissingNodesErrorStore()
-      const executionErrorStore = useExecutionErrorStore()
-      missingNodesStore.surfaceMissingNodes(
-        [{ type: 'NodeA', nodeId: '1', isReplaceable: false }],
-        () => executionErrorStore.showErrorOverlay()
-      )
-
-      expect(executionErrorStore.isErrorOverlayOpen).toBe(true)
-    })
-
-    it('does not open error overlay when ShowErrorsTab setting is false', () => {
-      mockShowErrorsTab.value = false
-      const missingNodesStore = useMissingNodesErrorStore()
-      const executionErrorStore = useExecutionErrorStore()
-      missingNodesStore.surfaceMissingNodes(
-        [{ type: 'NodeA', nodeId: '1', isReplaceable: false }],
-        () => executionErrorStore.showErrorOverlay()
-      )
-
-      expect(executionErrorStore.isErrorOverlayOpen).toBe(false)
-    })
-
-    it('deduplicates node types', () => {
-      const missingNodesStore = useMissingNodesErrorStore()
-      const executionErrorStore = useExecutionErrorStore()
-      missingNodesStore.surfaceMissingNodes(
-        [
-          { type: 'NodeA', nodeId: '1', isReplaceable: false },
-          { type: 'NodeA', nodeId: '1', isReplaceable: false },
-          { type: 'NodeB', nodeId: '2', isReplaceable: false }
-        ],
-        () => executionErrorStore.showErrorOverlay()
-      )
-
-      expect(missingNodesStore.missingNodesError?.nodeTypes).toHaveLength(2)
-    })
-  })
-
-  describe('removeMissingNodesByType', () => {
-    it('removes matching types from the missing nodes list', () => {
-      const store = useMissingNodesErrorStore()
-      store.setMissingNodeTypes([
-        { type: 'NodeA', nodeId: '1', isReplaceable: false },
-        { type: 'NodeB', nodeId: '2', isReplaceable: false },
-        { type: 'NodeC', nodeId: '3', isReplaceable: false }
-      ])
-
-      store.removeMissingNodesByType(['NodeA', 'NodeC'])
-
-      expect(store.missingNodesError?.nodeTypes).toHaveLength(1)
-      const remaining = store.missingNodesError?.nodeTypes[0]
-      expect(typeof remaining !== 'string' && remaining?.type).toBe('NodeB')
-    })
-
-    it('clears missingNodesError when all types are removed', () => {
-      const store = useMissingNodesErrorStore()
-      store.setMissingNodeTypes([
-        { type: 'NodeA', nodeId: '1', isReplaceable: false }
-      ])
-
-      store.removeMissingNodesByType(['NodeA'])
-
-      expect(store.missingNodesError).toBeNull()
-      expect(store.hasMissingNodes).toBe(false)
-    })
-
-    it('does nothing when missingNodesError is null', () => {
-      const store = useMissingNodesErrorStore()
-      expect(store.missingNodesError).toBeNull()
-
-      // Should not throw
-      store.removeMissingNodesByType(['NodeA'])
-      expect(store.missingNodesError).toBeNull()
-    })
-
-    it('does nothing when removing non-existent types', () => {
-      const store = useMissingNodesErrorStore()
-      store.setMissingNodeTypes([
-        { type: 'NodeA', nodeId: '1', isReplaceable: false }
-      ])
-
-      store.removeMissingNodesByType(['NonExistent'])
-
-      expect(store.missingNodesError?.nodeTypes).toHaveLength(1)
-    })
-
-    it('handles removing from string entries', () => {
-      const store = useMissingNodesErrorStore()
-      store.setMissingNodeTypes([
-        'StringNodeA',
-        'StringNodeB'
-      ] as MissingNodeType[])
-
-      store.removeMissingNodesByType(['StringNodeA'])
-
-      expect(store.missingNodesError?.nodeTypes).toHaveLength(1)
-    })
-  })
-})
-
 describe('executionErrorStore — node error operations', () => {
   beforeEach(() => {
     setActivePinia(createPinia())
